Update README.md
It uses weather API to fetch data from the server. It displays the temperature, humidity, cloudiness, and wind speed for the user's current location and any other city across the globe. Just enter the city name in the input field and the app will do the rest. If the city does not exist, it will give an error message -❌404